Split apart config into multiple files
The configuration files were getting huge (>400 LOC), so I split them up by structure.

Also added special websocket structure, and fixed Let's Encrypt bug.
