+++ src/config/configuration.go
@@ -5,16 +5,12 @@ Licensed under the MIT license.
 package config
 
 import (
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 
-	"github.com/dkumor/acmewrapper"
-	"github.com/gorilla/securecookie"
-	"github.com/nu7hatch/gouuid"
 	"github.com/tdewolff/minify"
 	"github.com/tdewolff/minify/js"
 
@@ -33,8 +29,8 @@ To see an explanation of the configuration, please see:
 
 http://connectordb.github.io/docs/config.html
 
-For a list of optinos and explanations of default values:
-https://github.com/connectordb/connectordb/blob/master/src/config/configuration.go
+For an explanation of default values:
+https://github.com/connectordb/connectordb/blob/master/src/config/defaultconfig.go
 	Look at NewConfiguration() which explains defaults.
 
 Particular configuration options:
@@ -100,7 +96,7 @@ type Configuration struct {
 	DatabaseDirectory string `json:"-"`
 }
 
-// Since we can't import the *actual* UserMaker from users (since that would give an import loop)
+// UserMaker: Since we can't import the *actual* UserMaker from users (since that would give an import loop)
 // we need to have our own version here - this version doesn't allow recusrive tree creation
 type UserMaker struct {
 	Name        string `json:"name"`        // The public username of the user
@@ -113,121 +109,6 @@ type UserMaker struct {
 	Public bool   `json:"public"`         // Whether the user is public or not
 
 	Password string `json:"password,omitempty"` // A hash of the user's password - it is never actually returned - the json params are used internally
-}
-
-// NewConfiguration generates a configuration with reasonable defaults for use in ConnectorDB
-func NewConfiguration() *Configuration {
-	redispassword, _ := uuid.NewV4()
-	natspassword, _ := uuid.NewV4()
-
-	sessionAuthKey := securecookie.GenerateRandomKey(64)
-	sessionEncKey := securecookie.GenerateRandomKey(32)
-
-	return &Configuration{
-		Version:     1,
-		Watch:       true,
-		Permissions: "default",
-		Redis: Service{
-			Hostname: "localhost",
-			Port:     6379,
-			Password: redispassword.String(),
-			Enabled:  true,
-		},
-		Nats: Service{
-			Hostname: "localhost",
-			Port:     4222,
-			Username: "connectordb",
-			Password: natspassword.String(),
-			Enabled:  true,
-		},
-		Sql: Service{
-			Hostname: "localhost",
-			Port:     52592,
-			//TODO: Have SQL accedd be auth'd
-			Enabled: true,
-		},
-
-		Frontend: Frontend{
-			Hostname: "0.0.0.0", // Host on all interfaces by default
-			Port:     8000,
-
-			Enabled: true,
-
-			// Sets up the session cookie keys that are used
-			CookieSession: CookieSession{
-				AuthKey:       base64.StdEncoding.EncodeToString(sessionAuthKey),
-				EncryptionKey: base64.StdEncoding.EncodeToString(sessionEncKey),
-				MaxAge:        60 * 60 * 24 * 30 * 4, //About 4 months is the default expiration time of a cookie
-			},
-
-			// By default, captcha is disabled
-			Captcha: Captcha{
-				Enabled: false,
-			},
-
-			// Set up the default TLS options
-			TLS: TLS{
-				Enabled: false,
-				Key:     "tls_key.pem",
-				Cert:    "tls_cert.crt",
-				ACME: ACME{
-					Server:       acmewrapper.DefaultServer,
-					PrivateKey:   "acme_privatekey.pem",
-					Registration: "acme_registration.json",
-					Domains:      []string{"example.com", "www.example.com"},
-					Enabled:      false,
-				},
-			},
-
-			// By default log query counts once a minute, and display server statistics
-			// once a day
-			QueryDisplayTimer: 60,
-			StatsDisplayTimer: 60 * 60 * 24,
-
-			// A limit of 10MB of data per insert is reasonable to me
-			InsertLimitBytes: 1024 * 1024 * 10,
-			// 1MB per websocket is also reasonable
-			WebsocketMessageLimitBytes: 1024 * 1024,
-
-			// The time to wait on a socket write in seconds
-			WebsocketWriteWait: 2,
-
-			// Websockets ping each other to keep the connection alive
-			// This sets the number of seconds between pings
-			WebsocketPongWait:   60,
-			WebsocketPingPeriod: 54,
-
-			// Websocket upgrader read/write buffer sizes
-			WebsocketReadBufferSize:  1024,
-			WebsocketWriteBufferSize: 1024,
-
-			// 3 messages should be enough... right?
-			WebsocketMessageBuffer: 3,
-
-			// Why not minify? Turning it off is useful for debugging - but users outnumber coders by a large margin.
-			Minify: true,
-		},
-
-		//The defaults to use for the batch and chunks
-		BatchSize: 250,
-		ChunkSize: 5,
-
-		UseCache:        true,
-		UserCacheSize:   1000,
-		DeviceCacheSize: 10000,
-		StreamCacheSize: 10000,
-
-		// This is the CONSTANT default. The database will explode if this is ever changed.
-		// You have been warned.
-		IDScramblePrime: 2147483423,
-
-		// No reason not to use bcrypt
-		PasswordHash: "bcrypt",
-
-		// Use the default settings.
-		PipeScript: psconfig.Default(),
-	}
-
 }
 
 // GetSqlConnectionString returns the string used to connect to postgres
